Obtain type URL via `TypeName`
