+++ client/src/main/java/org/spine3/type/KnownTypes.java
@@ -31,8 +31,6 @@ import com.google.protobuf.Api;
 import com.google.protobuf.BoolValue;
 import com.google.protobuf.BytesValue;
 import com.google.protobuf.DescriptorProtos;
-import com.google.protobuf.Descriptors;
-import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Descriptors.EnumDescriptor;
 import com.google.protobuf.DoubleValue;
 import com.google.protobuf.Duration;
@@ -66,17 +64,14 @@ import org.spine3.annotations.Internal;
 import org.spine3.type.error.UnknownTypeException;
 
 import javax.annotation.Nullable;
-import java.lang.reflect.InvocationTargetException;
 import java.util.Collection;
 import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 
-import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.collect.Maps.newHashMap;
 import static org.spine3.io.IoUtil.loadAllProperties;
-import static org.spine3.util.Exceptions.newIllegalArgumentException;
 import static org.spine3.util.Exceptions.newIllegalStateException;
 
 /**
@@ -95,9 +90,6 @@ public class KnownTypes {
      * <p>Is generated by Gradle during build process.
      */
     private static final String PROPS_FILE_PATH = "known_types.properties";
-
-    @SuppressWarnings("DuplicateStringLiteralInspection")
-    private static final String DESCRIPTOR_GETTER_NAME = "getDescriptor";
 
     private static final char PACKAGE_SEPARATOR = '.';
 
@@ -195,33 +187,6 @@ public class KnownTypes {
 
         final Set<TypeUrl> resultSet = ImmutableSet.copyOf(resultCollection);
         return resultSet;
-    }
-
-    /**
-     * Retrieve {@link Descriptors proto descriptor} from the type name.
-     *
-     * @param typeName <b>valid</b> name of the desired type
-     * @return {@link Descriptors proto descriptor} for given type
-     * @see TypeName
-     * @throws IllegalArgumentException if the name does not correspond to any known type
-     */
-    static Descriptor getDescriptorForType(String typeName) {
-        final TypeUrl typeUrl = getTypeUrl(typeName);
-        checkArgument(typeUrl != null, "Given type name is invalid");
-
-        final Class<?> cls = getJavaClass(typeUrl);
-
-        final Descriptor descriptor;
-        try {
-            final java.lang.reflect.Method descriptorGetter =
-                    cls.getDeclaredMethod(DESCRIPTOR_GETTER_NAME);
-            descriptor = (Descriptor) descriptorGetter.invoke(null);
-        } catch (NoSuchMethodException
-                | IllegalAccessException
-                | InvocationTargetException e) {
-            throw newIllegalArgumentException(e, "Type %s is not a protobuf Message", typeName);
-        }
-        return descriptor;
     }
 
     private static ImmutableMap<String, TypeUrl> buildTypeToUrlMap(BiMap<TypeUrl,
